Break apart the type of requests even further
We'll use multiple inheritance to simplify the commands.
