+++ src/apb/engine.py
@@ -2,7 +2,6 @@ import errno
 import os
 import uuid
 import base64
-import shutil
 import string
 import subprocess
 import json
@@ -16,6 +15,7 @@ from time import sleep
 from jinja2 import Environment, FileSystemLoader
 
 import request
+from broker_request.apb_list import Apb_List
 
 # Handle input in 2.x/3.x
 try:
@@ -96,41 +96,13 @@ class CmdRun(object):
             print("%s failure: {}".format(e) % cmd)
 
     def cmdrun_list(self):
-        return self.r.apb_list()
+        return Apb_List(self.args)
 
     def cmdrun_push(self):
         return self.r.apb_push()
 
-
-def load_dockerfile(df_path):
-    with open(df_path, 'r') as dockerfile:
-        return dockerfile.readlines()
-
-
-def load_makefile(apb_dict, params):
-    env = Environment(loader=FileSystemLoader(DAT_PATH), trim_blocks=True)
-    template = env.get_template(EX_MAKEFILE)
-
-    if not params:
-        params = []
-
-    return template.render(apb_dict=apb_dict, params=params)
-
-
-def load_example_specfile(apb_dict, params):
-    env = Environment(loader=FileSystemLoader(DAT_PATH), trim_blocks=True)
-    template = env.get_template(EX_SPEC_FILE)
-
-    if not params:
-        params = []
-
-    return template.render(apb_dict=apb_dict, params=params)
-
-
-def write_file(file_out, destination, force):
-    touch(destination, force)
-    with open(destination, 'w') as outfile:
-        outfile.write(''.join(file_out))
+    def cmdrun_init(self):
+        return self.r.apb_init()
 
 
 def insert_encoded_spec(dockerfile, encoded_spec_lines):
@@ -663,68 +635,6 @@ def build_apb(project, dockerfile=None, tag=None):
 
     print("Successfully built APB image: %s" % tag)
     return tag
-
-
-def cmdrun_init(**kwargs):
-    current_path = kwargs['base_path']
-    bindable = kwargs['bindable']
-    async = kwargs['async']
-    dockerhost = kwargs['dockerhost']
-    skip = {
-        'provision': kwargs['skip-provision'],
-        'deprovision': kwargs['skip-deprovision'],
-        'bind': kwargs['skip-bind'] or not kwargs['bindable'],
-        'unbind': kwargs['skip-unbind'] or not kwargs['bindable'],
-        'roles': kwargs['skip-roles']
-    }
-
-    apb_tag_arr = kwargs['tag'].split('/')
-    apb_name = apb_tag_arr[-1]
-    app_org = apb_tag_arr[0]
-    if apb_name.lower().endswith("-apb"):
-        app_name = apb_name[:-4]
-    else:
-        app_name = apb_name
-
-    description = "This is a sample application generated by apb init"
-
-    apb_dict = {
-        'name': apb_name,
-        'app_name': app_name,
-        'app_org': app_org,
-        'description': description,
-        'bindable': bindable,
-        'async': async,
-        'dockerhost': dockerhost
-    }
-
-    project = os.path.join(current_path, apb_name)
-
-    if os.path.exists(project):
-        if not kwargs['force']:
-            raise Exception('ERROR: Project directory: [%s] found and force option not specified' % project)
-        shutil.rmtree(project)
-
-    print("Initializing %s for an APB." % project)
-
-    os.mkdir(project)
-
-    spec_path = os.path.join(project, SPEC_FILE)
-    dockerfile_path = os.path.join(os.path.join(project, DOCKERFILE))
-    makefile_path = os.path.join(os.path.join(project, MAKEFILE))
-
-    specfile_out = load_example_specfile(apb_dict, [])
-    write_file(specfile_out, spec_path, kwargs['force'])
-
-    dockerfile_out = load_dockerfile(EX_DOCKERFILE_PATH)
-    write_file(dockerfile_out, dockerfile_path, kwargs['force'])
-
-    makefile_out = load_makefile(apb_dict, [])
-    write_file(makefile_out, makefile_path, kwargs['force'])
-
-    generate_playbook_files(project, skip, apb_dict)
-    print("Successfully initialized project directory at: %s" % project)
-    print("Please run *apb prepare* inside of this directory after editing files.")
 
 
 def cmdrun_prepare(**kwargs):
